Adição de codigo na main, fuções gerar chave e gerr primos funcionando

+++ CRIPTO_RSA/GerarChave.cs
@@ -1,38 +1,40 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System.IO;
+using System.Numerics;
 
 namespace CRIPTO_RSA
 {
-
     internal class GerarChave
     {
-     public ulong Mdc(ulong a, ulong b)
+        public BigInteger Mdc(BigInteger a, BigInteger b)
         {
-            while (b!=0)
+            while (b != 0)
             {
-                ulong temp = a % b;
+                BigInteger temp = a % b;
                 a = b;
                 b = temp;
             }
             return a;
         }
-      public bool Checkprimo(ulong n)
+
+        public bool Checkprimo(BigInteger n)
         {
-            if (n < 2) return false;
-            ulong maiordiv = (ulong)Math.Sqrt(n);
-            for (ulong i = 2; i <= maiordiv; i++)
+            if (n < 2)
+                return false;
+            BigInteger maiordiv = (BigInteger)Math.Sqrt((double)n);
+            for (BigInteger i = 2; i <= maiordiv; i++)
             {
-                if (n % i == 0) return false;
+                if (n % i == 0)
+                    return false;
             }
-           return true;
+            return true;
         }
-      public void Genchave(ulong p, ulong q, ulong e)
+
+
+        public void Genchave(BigInteger p, BigInteger q, BigInteger e)
         {
-            ulong n = p * q;
-            ulong phi = (p - 1) * (q - 1);
+            BigInteger n = p * q;
+            BigInteger phi = (p - 1) * (q - 1);
 
             using (StreamWriter writer = new StreamWriter("chave_publica.txt"))
             {
